move complex title to Marketplace component
